# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Campus Library student resource portal
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type MaterialStatus = typeof MaterialStatus[keyof typeof MaterialStatus];


export const MaterialStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Material {
  id: number;
  title: string;
  description: string;
  subject: string;
  courseCode: string;
  category: string;
  semester: string;
  year: number;
  fileName: string;
  fileSize: number;
  fileType: string;
  /** @nullable */
  objectPath: string | null;
  uploadedBy: string;
  uploadedByName: string;
  status: MaterialStatus;
  downloads: number;
  createdAt: string;
  /** @nullable */
  reviewNote: string | null;
}

export interface MaterialInput {
  /** @minLength 1 */
  title: string;
  description: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  courseCode: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  semester: string;
  year: number;
  /** @minLength 1 */
  fileName: string;
  /** @minimum 1 */
  fileSize: number;
  /** @minLength 1 */
  fileType: string;
  /** @nullable */
  objectPath?: string | null;
  /** @minLength 1 */
  uploadedBy: string;
  /** @minLength 1 */
  uploadedByName: string;
}

export interface ReviewInput {
  reviewNote: string;
}

export interface SubjectSummary {
  subject: string;
  count: number;
}

export interface DashboardStats {
  totalMaterials: number;
  totalSubjects: number;
  pendingSubmissions: number;
  totalDownloads: number;
  popularSubjects: SubjectSummary[];
}

export interface UploadRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListMaterialsParams = {
search?: string;
subject?: string;
category?: string;
semester?: string;
sort?: ListMaterialsSort;
};

export type ListMaterialsSort = typeof ListMaterialsSort[keyof typeof ListMaterialsSort];


export const ListMaterialsSort = {
  recent: 'recent',
  popular: 'popular',
} as const;

export type ListAdminSubmissionsParams = {
status?: ListAdminSubmissionsStatus;
};

export type ListAdminSubmissionsStatus = typeof ListAdminSubmissionsStatus[keyof typeof ListAdminSubmissionsStatus];


export const ListAdminSubmissionsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  all: 'all',
} as const;