# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Campus Library student resource portal
 * OpenAPI spec version: 0.1.0
 */
import type { SubjectSummary } from './subjectSummary';

export interface DashboardStats {
  totalMaterials: number;
  totalSubjects: number;
  pendingSubmissions: number;
  totalDownloads: number;
  popularSubjects: SubjectSummary[];
}