# lib/api-zod/src/generated/types/material.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Campus Library student resource portal
 * OpenAPI spec version: 0.1.0
 */
import type { MaterialStatus } from './materialStatus';

export interface Material {
  id: number;
  title: string;
  description: string;
  subject: string;
  courseCode: string;
  category: string;
  semester: string;
  year: number;
  fileName: string;
  fileSize: number;
  fileType: string;
  /** @nullable */
  objectPath: string | null;
  uploadedBy: string;
  uploadedByName: string;
  status: MaterialStatus;
  downloads: number;
  createdAt: Date;
  /** @nullable */
  reviewNote: string | null;
}