# lib/api-zod/src/generated/types/materialInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Campus Library student resource portal
 * OpenAPI spec version: 0.1.0
 */

export interface MaterialInput {
  /** @minLength 1 */
  title: string;
  description: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  courseCode: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  semester: string;
  year: number;
  /** @minLength 1 */
  fileName: string;
  /** @minimum 1 */
  fileSize: number;
  /** @minLength 1 */
  fileType: string;
  /** @nullable */
  objectPath?: string | null;
  /** @minLength 1 */
  uploadedBy: string;
  /** @minLength 1 */
  uploadedByName: string;
}