# lib/api-zod/src/generated/types/uploadResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Campus Library student resource portal
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}